UI changes
Changes to main, speakers and about pages. Also did refactoring and removed livestream icons, layouts and references.
