+++ .idea/assetWizardSettings.xml
@@ -23,7 +23,7 @@
                                           <PersistentState>
                                             <option name="values">
                                               <map>
-                                                <entry key="url" value="jar:file:/C:/Program%20Files/Android/Android%20Studio/plugins/android/lib/android.jar!/images/material_design_icons/social/ic_sentiment_very_dissatisfied_black_24dp.xml" />
+                                                <entry key="url" value="jar:file:/C:/Program%20Files/Android/Android%20Studio/plugins/android/lib/android.jar!/images/material_design_icons/navigation/ic_close_black_24dp.xml" />
                                               </map>
                                             </option>
                                           </PersistentState>
@@ -33,8 +33,8 @@
                                   </option>
                                   <option name="values">
                                     <map>
-                                      <entry key="outputName" value="ic_sentiment_one_dark" />
-                                      <entry key="theme" value="CUSTOM" />
+                                      <entry key="outputName" value="ic_close" />
+                                      <entry key="theme" value="HOLO_DARK" />
                                       <entry key="themeColor" value="f5f5f5" />
                                     </map>
                                   </option>
