+++ app/src/main/java/com/example/kmike/gdgkisumupreview/ResourcesActivity.java
@@ -2,6 +2,7 @@ package com.example.kmike.gdgkisumupreview;
 
 import android.os.Bundle;
 import android.preference.PreferenceManager;
+import android.support.design.widget.CoordinatorLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
@@ -24,16 +25,37 @@ public class ResourcesActivity extends AppCompatActivity {
         setContentView(R.layout.activity_resources);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
+        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
 
-        FloatingActionButton fab = findViewById(R.id.fab);
+        final FloatingActionButton fab = findViewById(R.id.fab);
+        final Snackbar snackbar = Snackbar.make(findViewById(R.id.layoutResources), "Please refresh before requesting a resource.", Snackbar.LENGTH_INDEFINITE)
+                .setAction("Request", new View.OnClickListener() {
+                    @Override
+                    public void onClick(View view) {
+                        Toast.makeText(ResourcesActivity.this, "Resource requested", Toast.LENGTH_SHORT).show();
+                    }
+                });
+        snackbar.addCallback(new Snackbar.Callback(){
+            @Override
+            public void onDismissed(Snackbar transientBottomBar, int event) {
+                fab.setImageDrawable(getResources().getDrawable(R.drawable.ic_request));
+            }
+
+            @Override
+            public void onShown(Snackbar sb) {
+                fab.setImageDrawable(getResources().getDrawable(R.drawable.ic_close));
+            }
+        });
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
+                if (snackbar.isShown()){
+                    snackbar.dismiss();
+                } else{
+                    snackbar.show();
+                }
             }
         });
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
     }
 
     @Override
