+++ app/src/main/res/drawable/circle.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="ring"
+    android:innerRadius="0dp"
+    android:thicknessRatio="2.0"
+    android:useLevel="false">
+
+    <solid android:color="@color/transparent" />
+
+    <stroke android:color="?attr/color_background"
+        android:width="10dp"/>
+</shape>
