+++ app/src/main/res/layout/item_resource.xml
@@ -22,7 +22,6 @@
             android:orientation="vertical">
 
             <TextView
-                android:id="@+id/tvResourceTitle"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:fontFamily="@font/nunito_light"
@@ -30,14 +29,12 @@
                 android:text="Git Resource Pack"/>
 
             <TextView
-                android:id="@+id/tvResourceEvent"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:textSize="16sp"
                 android:text="Mentioned in How to Git"/>
 
             <TextView
-                android:id="@+id/tvResourceDetails"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:textSize="16sp"
@@ -48,7 +45,6 @@
                 android:layout_height="wrap_content">
 
                 <Button
-                    android:id="@+id/btnCopy"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_alignParentStart="true"
@@ -59,7 +55,6 @@
                     android:text="@string/copy_link"/>
 
                 <Button
-                    android:id="@+id/btnOpen"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_alignParentEnd="true"
