+++ app/src/main/res/values/strings.xml
@@ -8,32 +8,32 @@
     <string name="schedule">Schedule</string>
     <string name="schedule_desc">What\'s going on, when and where?</string>
     <string name="schedule_icon">Schedule Icon</string>
+    <string name="title_activity_schedule">Schedule</string>
+    <string name="add_to_calendar">Add to calendar</string>
 
-    <string name="livestream">LiveStream</string>
-    <string name="livestream_desc">Watch from wherever</string>
-    <string name="livestream_icon">Live stream icon</string>
+    <string name="speakers">Speakers</string>
+    <string name="speakers_desc">Who will be talking about what?</string>
+    <string name="speakers_icon">Speakers icon</string>
+    <string name="title_activity_speakers">Speakers</string>
 
     <string name="resources">Resources</string>
     <string name="resources_desc">Other helpful stuff</string>
     <string name="resources_icon">Resources icon</string>
+    <string name="title_activity_resources">Resources</string>
+    <string name="copy_link">Copy Link</string>
+    <string name="open_resource">Open Resource</string>
 
     <string name="about_us">About Us</string>
     <string name="about_us_desc">About GDG Kisumu</string>
     <string name="about_icon">About icon</string>
-
     <string name="title_activity_about">About Us</string>
     <string name="we_need_your_feedback">We need your feedback!</string>
     <string name="str_feedback">We need your feedback</string>
     <string name="feedback">Feedback</string>
 
-    <string name="title_activity_schedule">Schedule</string>
-    <string name="title_activity_resources">Resources</string>
-    <string name="add_to_calendar">Add to calendar</string>
-    <string name="copy_stream_link">copy stream link</string>
-    <string name="title_activity_livestream">LivestreamActivity</string>
-    <string name="select_a_room">Select a room</string>
     <string name="refresh">Refresh</string>
-    <string name="copy_link">Copy Link</string>
-    <string name="open_resource">Open Resource</string>
+    <string name="str_host_bio_test">This speaker has obviously either graduated from some well-known university with high grades or has some tragic sob-story that ends in a path to self-discovery. He is a master in an emerging technological field.</string>
+    <string name="gdg_about_pt_one">GDGs are local groups of developers who are specifically interested in Google products and APIs. Each local group is called a GDG chapter and can host a variety of technical activities for developers - just a few people getting together to watch our latest videos,to large gatherings with demos annd tech talks,to hackathons. GDG Kisumu is the Kisumu chapter, we also have Eldoret, Mombasa and Nairobi chapters. If you are a student then there is DSC in the universities.</string>
+    <string name="gdg_about_pt_two">We have various events in the course of the year majorly;\n● International Women\'s Day\n● Google IO extended\n● Google cloud\n● GDG Developer Festival commonly known as DEVFEST\nWe also occasionally have study jams.</string>
 
 </resources>
