+++ app/src/main/res/values/styles.xml
@@ -11,6 +11,7 @@
         <item name="android:textColorSecondary">@color/colorTextSecondary</item>
         <item name="android:textColorPrimaryInverse">@color/colorTextPrimaryInverse</item>
         <item name="android:textColorSecondaryInverse">@color/colorTextSecondaryInverse</item>
+        <item name="color_background">@color/colorPrimary</item>
         <item name="ic_sentiment_one">@drawable/ic_sentiment_one_light</item>
         <item name="ic_sentiment_two">@drawable/ic_sentiment_two_light</item>
         <item name="ic_sentiment_three">@drawable/ic_sentiment_three_light</item>
@@ -21,13 +22,14 @@
     <style name="AppTheme.Dark" parent="Theme.AppCompat.NoActionBar">
         <item name="colorPrimary">@color/colorPrimaryInverse</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDarkInverse</item>
-        <item name="colorAccent">@color/colorAccent</item>
+        <item name="colorAccent">@color/colorAccentInverse</item>
         <item name="android:windowBackground">@color/colorPrimaryInverse</item>
         <item name="android:fontFamily">@font/nunito_extra_light</item>
         <item name="android:textColorPrimary">@color/colorTextPrimaryInverse</item>
         <item name="android:textColorSecondary">@color/colorTextSecondaryInverse</item>
         <item name="android:textColorPrimaryInverse">@color/colorTextPrimary</item>
         <item name="android:textColorSecondaryInverse">@color/colorTextSecondary</item>
+        <item name="color_background">@color/colorPrimaryInverse</item>
         <item name="ic_sentiment_one">@drawable/ic_sentiment_one_dark</item>
         <item name="ic_sentiment_two">@drawable/ic_sentiment_two_dark</item>
         <item name="ic_sentiment_three">@drawable/ic_sentiment_three_dark</item>
@@ -73,14 +75,4 @@
 
     <style name="AppTheme.PopupOverlay" parent="ThemeOverlay.AppCompat" />
 
-    <style name="BottomSheet" parent="@style/Widget.Design.BottomSheet.Modal">
-        <item name="android:background">?android:windowBackground</item>
-    </style>
-
-    <style name="BaseBottomSheetDialog" parent="Theme.Design.BottomSheetDialog">
-        <item name="android:windowIsFloating">false</item>
-        <item name="bottomSheetStyle">@style/BottomSheet</item>
-    </style>
-
-    <style name="BottomSheetDialogTheme" parent="BaseBottomSheetDialog" />
 </resources>
